requirements.txt
Flask==2.0.1
mysql-connector-python==8.0.26
apscheduler==3.7.0
pytz==2021.1
